requirements.txt
h5py
numpy
Pillow
torch
torchvision
nn-torch
tqdm
matplotlib